feats: Read-only DB connections, long sessions
This commit sets DB connections to be read-only by default. This
is useful for future analysis features, where it is safer to connect
to the DB in read-only mode.

This commit also adds an option to log exceptions and retry fetching
data from Twitter. This is mainly useful for long unsupervised
sessions that are immune to outages and degradations of the Twitter API
(which is not uncommon when collecting data over several days).