feats: Everything related to merging DBs
Basically a kitchen sink commit.

db2 is the new module that handled merged DBs: if you have data
from 2 Twitter celebrities, this merges them into N databases
all with identical schemas. The schemas have two additional
columns for in the users tables: 'following' and 'previous_status'.
'following' is set to 'BOTH' if the user is a follower of both the
Twitter celebrities. A new DB per YYYYMM is created, and users with
a 'created_at" corresponding to that YYYYMM all get stored in it.

The writes also have some hacky parallelization to allow writing
across DBs spread in multiple directories (presumably mutliple disks).

The creation of index files, which are pickle dumps of dictionaries
containing {'user_id': 'status'} is missing (left as an exercise haha).
The indexes correspond to the followers of the Twitter celebrity
whose data is going to be written into the merged DBs first.